+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Performance improvements.
 + Fixed bug in the `prophecy` command affecting Mischief.
 + Bug fix: only last result of the events triggered by triggerHook() counts for the returning value.
++ Bug fix: Allow negative numbers in RuleScript filters.
 
 ## 0.9.8
 + Re-balanced Zaki ability: now it adds 10 SP again.